fAlterado tipo atr altura para int

+++ src/main/java/com/viniciusx/proj_imc_on/model/Imc.java
@@ -21,7 +21,7 @@ public class Imc {
     private Long id;
     private String nome;
     private double peso;
-    private double altura;
+    private int altura;
     private String sexo;
     private double valorImc;
     @JsonFormat(pattern = "dd/MM/yyyy HH:mm:ss")
